requirements.txt
# Data Ingestion & API
requests
geopy
pandas

# Transformation & Warehouse
dbt-duckdb
duckdb

# Utility
python-dotenv